refactor: split Service logic
